+++ app/static/css/home.css
@@ -1,0 +1,21 @@
+/* static/css/home.css */
+.search-form {
+    margin-top: 20px;
+}
+.search-form input[type="text"] {
+    padding: 8px;
+    width: 200px;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+}
+.search-form input[type="submit"] {
+    padding: 8px 16px;
+    border: none;
+    background-color: #28a745;
+    color: white;
+    border-radius: 4px;
+    cursor: pointer;
+}
+.search-form input[type="submit"]:hover {
+    background-color: #218838;
+}
