+++ app/templates/base.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <meta charset="utf-8">
+    <title>{% block title %}Library{% endblock %}</title>
+    <link rel="stylesheet"
+          type="text/css"
+          href="{{ url_for('static', filename='css/main.css') }}">
+    {% block extra_css %}{% endblock %}
+</head>
+
+<body>
+    <h1>도서관</h1>
+    <nav>
+        <a href="{{ url_for('home') }}">홈</a> | <a href="{{ url_for('login') }}">로그인</a>
+    </nav>
+    <hr>
+
+    {% block content %}{% endblock %}
+</body>
+
+</html>
