+++ app/templates/home.html
@@ -1,0 +1,19 @@
+<!-- app/templates/home.html -->
+{% extends "base.html" %}
+
+{% block title %}Home - Library{% endblock %}
+
+{% block content %}
+    <h2>책 검색</h2>
+    <form action="{{ url_for('search') }}" method="POST" class="search-form">
+        <input type="text" name="query" placeholder="책 제목을 입력하세요" required>
+        <input type="submit" value="검색">
+    </form>
+    
+    <!-- Error Message Display -->
+    {% if error %}
+        <div class="error-message">
+            <p>{{ error }}</p>
+        </div>
+    {% endif %}
+{% endblock %}
