+++ app/templates/login.html
@@ -1,0 +1,8 @@
+<!DOCTYPE html>
+<head>
+    <h1></h1>
+</head>
+<body>
+    <h1>Login</h1>
+</body>
+</html>
