+++ main.py
@@ -1,26 +1,6 @@
-import pandas as pd
-import sqlite3
-import re
-from flask import Flask, jsonify, render_template
-from book_rental_system import LibraryDatabase
+from app import create_app
 
-app = Flask(__name__)
+app = create_app()
 
-@app.route('/')
-def home():
-    return render_template('home.html')
-
-@app.route('/search')
-def search():
-    return render_template('search.html')
-
-@app.route('/login')
-def log_in():
-    return render_template('login.html')
-  
-if __name__ == "__main__":
-    # First, create tables and load data from the CSV file
-    
-    LibraryDatabase(db_name="library.db", csv_file="library.csv")
-    
-    app.run(debug=True)
+if __name__ == '__main__':
+    app.run(debug=True, host='127.0.0.1', port=5000)
