--- templates/home.html
@@ -1,15 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-    <head>
-        <title>Welcome To the libary</title>
-        <meta charset="utf-8">
-        <style>
-            h1{
-                color:aqua
-            }
-        </style>
-    </head>
-    <body>
-        <h1>Welcome</h1>
-    </body>
-</html>
