--- templates/login.html
@@ -1,8 +1,0 @@
-!<!DOCTYPE html>
-<head>
-    <h1></h1>
-</head>
-<body>
-    <h1></h1>
-</body>
-</html>
