--- templates/search.html
@@ -1,8 +1,0 @@
-<!DOCTYPE html>
-<head>
-    <title></title>
-</head>
-<body>
-    <h1>Hi</h1>
-</body>
-</html>
